finally figured out python packaging... I think
