+++ ansifier/__init__.py
@@ -1,0 +1,2 @@
+from .ansifier import ImageFilePrinter
+from .config import *
