+++ pyproject.toml
@@ -2,22 +2,14 @@
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
-[tool.hatch.build.targets.sdist]
-include = [
-  "/*.py",
-  "/config",
-  "/tests",
-  "/images-examples"
-]
-exclude = [
-]
-
 [tool.hatch.build.targets.wheel]
-packages = ["."]
+include = [
+  "/ansifier/",
+]
 
 [project]
 name = "ansifier"
-version = "0.0.2"
+version = "0.0.3"
 authors = [
   { name="Amelia Miner", email="ameliamminer@gmail.com" },
 ]
